requirements.txt
pyspark==3.5.1
delta-spark==2.4.0
